+++ deps.tsx
@@ -4,10 +4,12 @@
 
 //React
 export * as ReactDOM from "https://esm.sh/react-dom@18.2.0"
-export { default as ReactDOMServer } from 'https://dev.jspm.io/react-dom@16.14.0/server';
+// export { default as ReactDOMServer } from 'https://dev.jspm.io/react-dom@16.14.0/server';
 import * as React from "https://esm.sh/react@18.2.0"
 export { React }
 
 //Twind
 import { setup, tw } from "https://esm.sh/twind@0.16.16";
 import { getStyleTag, virtualSheet } from "https://esm.sh/twind@0.16.16/sheets";
+
+// import ReactDOMServer from "https://dev.jspm.io/react-dom@16.13.1/server";
