+++ index.html
@@ -10,7 +10,7 @@
     <title>Squeezed</title>
 </head>
 <body>
-<noscript>You need to enable JavaScript to run this app.</noscript>
+<noscript>In Index Html</noscript>
 <div id="root"></div>
 </body>
 <script src="./static/js/main.js"></script>
